+++ NEWS.md
@@ -3,5 +3,6 @@
 
 # CURRENT: 0.0.3
 ### Breaking Changes
-- `background_color()` renamed to `newpage()`.
+- `background()` renamed to `newpage()`.
 	- default parameter `auto_refresh` set to `True`.
+- `background` argument in all functions renamed `background`